Feat/menu create 38 (#49)
* Refactor : 가게, 메뉴 코드 리팩토링

* Feat : 메뉴 등록 외부 AI 연동 하여 생성 기능 개발

* Refactor : 도메인 계층 Repository 삭제

* Fix : MenuRepository 변경으로 컴파일 오류 부분 수정

+++ src/main/java/com/driven/dm/menu/application/service/MenuService.java
@@ -1,11 +1,13 @@
 package com.driven.dm.menu.application.service;
 
+import com.driven.dm.ai.application.service.AiService;
+import com.driven.dm.ai.presentation.dto.response.AiCallResponseDto;
 import com.driven.dm.global.config.security.SecurityUser;
 import com.driven.dm.global.exception.AppException;
 import com.driven.dm.menu.application.exception.MenuErrorCode;
 import com.driven.dm.menu.domain.entity.Menu;
 import com.driven.dm.menu.domain.entity.MenuStatus;
-import com.driven.dm.menu.domain.repository.MenuRepository;
+import com.driven.dm.menu.infrastructure.repository.MenuRepository;
 import com.driven.dm.menu.presentation.dto.request.MenuCreateRequest;
 import com.driven.dm.menu.presentation.dto.request.MenuUpdateRequest;
 import com.driven.dm.menu.presentation.dto.response.MenuCreateResponse;
@@ -15,7 +17,8 @@ import com.driven.dm.menu.presentation.dto.response.MenuShopResponse.MenuRespons
 import com.driven.dm.menu.presentation.dto.response.MenuUpdateResponse;
 import com.driven.dm.shop.application.exception.ShopErrorCode;
 import com.driven.dm.shop.domain.entity.Shop;
-import com.driven.dm.shop.domain.repository.ShopRepository;
+import com.driven.dm.shop.domain.entity.ShopStatus;
+import com.driven.dm.shop.infrastructure.repository.ShopJpaRepository;
 import com.driven.dm.user.application.exception.UserErrorCode;
 import com.driven.dm.user.domain.entity.User;
 import com.driven.dm.user.domain.entity.UserRole;
@@ -31,29 +34,53 @@ import org.springframework.transaction.annotation.Transactional;
 public class MenuService {
 
     private final UserRepository userRepository;
-    private final ShopRepository shopRepository;
+    private final ShopJpaRepository shopRepository;
     private final MenuRepository menuRepository;
+    private final AiService aiService;
 
     @Transactional
-    public MenuCreateResponse createMenu(UUID id, MenuCreateRequest menuCreateRequest) {
-        Shop shop = getShop(id);
-        User user = userRepository.findById(shop.getOwner().getId()).orElseThrow(
-            () -> new AppException(UserErrorCode.USER_NOT_FOUND)
-        );
+    public MenuCreateResponse createMenu(UUID shop_id, SecurityUser securityUser, MenuCreateRequest menuCreateRequest) {
 
-        if (isOwner(user, shop)) {
+        String menuDescription;
+        boolean isAi;
+
+        Shop shop = getShop(shop_id);
+        User user = getUser(securityUser);
+
+        if (!isOwner(user, shop)
+            || user.getRole().equals(UserRole.MANAGER)
+            || user.getRole().equals(UserRole.MASTER)) {
             throw new AppException(ShopErrorCode.SHOP_NOT_OWNER);
         }
 
-        Menu menu = Menu.of(shop, menuCreateRequest.getMenuname(),
-            menuCreateRequest.getMenuprice());
-        Menu saveMenu = menuRepository.createMenu(menu).orElseThrow(
-            () -> new AppException(MenuErrorCode.MENU_SAVE_FAIL)
-        );
+        if (menuCreateRequest.isGenerateDescription()
+            && !menuCreateRequest.getKeyword().isEmpty()
+            && !menuCreateRequest.getFeatures().isEmpty()) {
+            AiCallResponseDto aiCallResponseDto = aiService.generateMenuDescription(
+                user.getId(),
+                menuCreateRequest.getMenuName(),
+                menuCreateRequest.getKeyword(),
+                menuCreateRequest.getFeatures());
+            menuDescription = aiCallResponseDto.getOutputText();
+            isAi = true;
+        } else {
+            menuDescription = menuCreateRequest.getMenuDescription();
+            isAi = false;
+        }
+
+        Menu menu =
+            Menu.of(shop,
+                menuCreateRequest.getMenuName(),
+                menuDescription,
+                menuCreateRequest.getMenuPrice(),
+                menuCreateRequest.getKeyword(), isAi);
+
+        Menu saveMenu = menuRepository.save(menu);
 
         return MenuCreateResponse.builder()
-            .menuId(saveMenu.getId())
             .menuName(saveMenu.getMenuName())
+            .menuDescription(saveMenu.getMenuDescription())
+            .menuKeyword(saveMenu.getMenuKeyword())
             .menuPrice(saveMenu.getMenuPrice())
             .build();
     }
@@ -61,7 +88,7 @@ public class MenuService {
     @Transactional(readOnly = true)
     public List<MenuListResponse> menuList(SecurityUser securityUser) {
 
-        List<Menu> menus = menuRepository.selectAll();
+        List<Menu> menus = menuRepository.findAll();
 
         boolean isPrivileged =
             securityUser.getRole().equals(UserRole.MASTER)
@@ -69,8 +96,9 @@ public class MenuService {
 
         return menus.stream()
             .filter(menu ->
-                isPrivileged ? (menu.getStatus().equals(MenuStatus.ACTIVE) || menu.getStatus().equals(MenuStatus.HIDDEN))
-                            : menu.getStatus().equals(MenuStatus.ACTIVE))
+                isPrivileged ? (menu.getStatus().equals(MenuStatus.ACTIVE) || menu.getStatus()
+                    .equals(MenuStatus.HIDDEN))
+                    : menu.getStatus().equals(MenuStatus.ACTIVE))
             .map(MenuListResponse::from)
             .toList();
     }
@@ -78,20 +106,41 @@ public class MenuService {
     @Transactional
     public MenuUpdateResponse updateMenu(UUID id, SecurityUser securityUser,
         MenuUpdateRequest menuUpdateRequest) {
+
         User user = getUser(securityUser);
-        Menu menu = menuRepository.selectMenu(id).orElseThrow(
+        Menu menu = menuRepository.findById(id).orElseThrow(
             () -> new AppException(MenuErrorCode.MENU_NOT_FOUND)
         );
+
+        if (menu.getStatus().equals(MenuStatus.DELETED)) {
+            throw new AppException(MenuErrorCode.MENU_NOT_FOUND);
+        }
 
         if (!isOwner(user, menu.getShop())) {
             throw new AppException(ShopErrorCode.SHOP_NOT_OWNER);
         }
 
-        menu.changeMenuName(menuUpdateRequest.getMenuName());
-        menu.changeMenuPrice(menuUpdateRequest.getMenuPrice());
-        menu.changeStatus(menuUpdateRequest.getMenuStatus());
+        if (!menuUpdateRequest.getMenuName().isEmpty()) {
+            menu.changeMenuName(menuUpdateRequest.getMenuName());
+        }
 
-        Menu updatedMenu = menuRepository.updateMenu(menu);
+        if (!menuUpdateRequest.getMenuPrice().isEmpty()) {
+            menu.changeMenuPrice(Long.parseLong(menuUpdateRequest.getMenuPrice()));
+        }
+
+        if (!menuUpdateRequest.getMenuDescription().isEmpty()) {
+            menu.changeMenuDescription(menuUpdateRequest.getMenuDescription());
+        }
+
+        if (!menuUpdateRequest.getMenuKeyword().isEmpty()) {
+            menu.changeMenuKeyword(menuUpdateRequest.getMenuKeyword());
+        }
+
+        if (!menuUpdateRequest.getMenuStatus().isEmpty()) {
+            menu.changeStatus(menuUpdateRequest.getMenuStatus());
+        }
+
+        Menu updatedMenu = menuRepository.save(menu);
 
         return MenuUpdateResponse.builder()
             .menuName(updatedMenu.getMenuName())
@@ -104,37 +153,48 @@ public class MenuService {
         Menu menu = getMenu(id);
         User user = getUser(securityUser);
 
+        if (menu.getStatus().equals(MenuStatus.DELETED)) {
+            throw new AppException(MenuErrorCode.MENU_NOT_FOUND);
+        }
+
         if (!isOwner(user, menu.getShop())) {
             throw new AppException(ShopErrorCode.SHOP_NOT_OWNER);
         }
 
         menu.deleteMenu();
-        menuRepository.deleteMenu(menu);
+        menuRepository.save(menu);
     }
 
     @Transactional(readOnly = true)
-    public MenuShopResponse shopMenuList(UUID id, SecurityUser securityUser) {
+    public MenuShopResponse shopMenuList(UUID shop_id, SecurityUser securityUser) {
 
         User user = getUser(securityUser);
-        Shop shop = shopRepository.findByIdWithMenus(id).orElseThrow(
+        Shop shop = shopRepository.findById(shop_id).orElseThrow(
             () -> new AppException(ShopErrorCode.SHOP_NOT_FOUND)
         );
 
+        if (shop.getStatus().equals(ShopStatus.DELETED)) {
+            throw new AppException(ShopErrorCode.SHOP_NOT_FOUND);
+        }
+
         boolean isPrivileged =
             isOwner(user, shop)
-            || user.getRole() == UserRole.MANAGER
-            || user.getRole() == UserRole.MASTER;
+                || user.getRole() == UserRole.MANAGER
+                || user.getRole() == UserRole.MASTER;
 
         List<MenuResponse> menuResponses = shop.getMenu().stream()
             .filter(menu -> {
                 if (isPrivileged) {
-                    return menu.getStatus() == MenuStatus.ACTIVE || menu.getStatus() == MenuStatus.HIDDEN;
+                    return menu.getStatus() == MenuStatus.ACTIVE
+                        || menu.getStatus() == MenuStatus.HIDDEN;
                 } else {
                     return menu.getStatus() == MenuStatus.ACTIVE;
                 }
             })
             .map(menu -> MenuResponse.builder()
                 .menuName(menu.getMenuName())
+                .menuDescription(menu.getMenuDescription())
+                .menuKeyword(menu.getMenuKeyword())
                 .menuPrice(menu.getMenuPrice())
                 .build())
             .toList();
@@ -147,13 +207,13 @@ public class MenuService {
 
     private Menu getMenu(UUID id) {
 
-        return menuRepository.selectMenu(id).orElseThrow(
+        return menuRepository.findById(id).orElseThrow(
             () -> new AppException(MenuErrorCode.MENU_NOT_FOUND)
         );
     }
 
     private Shop getShop(UUID id) {
-        Shop shop = shopRepository.selectShop(id).orElseThrow(
+        Shop shop = shopRepository.findById(id).orElseThrow(
             () -> new AppException(ShopErrorCode.SHOP_NOT_FOUND)
         );
